Move dispatchers to simpleflow.dispatch
This avoids a cyclic dependency import problem in the next commit.
